requirements.txt
Pillow>=9.0.0
apng>=0.3.4